+++ Banking/View/WindowContent.xaml
@@ -7,6 +7,9 @@
              xmlns:CornUI="clr-namespace:CornUI.Controls.Normal;assembly=CornUI"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-    <CornUI:CButton Width="100" Height="50" CornerRadius="10" BorderThickness="2, 2, 2, 2"
+    <Grid>
+        <CornUI:CButton Width="100" Height="50" CornerRadius="10" BorderThickness="2, 2, 2, 2"
                     ShadowDirection="270" IconSize="40" FontSize="29" VerticalAlignment="Bottom" HorizontalAlignment="Left"/>
+        <CornUI:CTextbox Width="100" Height="50"/>
+    </Grid>
 </UserControl>
