+++ CornUI/Controls/Normal/CTextbox.xaml
@@ -1,0 +1,32 @@
+﻿<local:CornControl x:Class="CornUI.Controls.Normal.CTextbox"
+                   x:Name="root"
+                   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+                   xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+                   xmlns:local="clr-namespace:CornUI.Controls.Normal"
+                   mc:Ignorable="d" 
+                   d:DesignHeight="50" d:DesignWidth="100"
+                   CornerRadius="10">
+    <Grid>
+        <TextBox x:Name="textBox"
+                 BorderBrush="{Binding ElementName=root, Path=CurrentBorderColor}"
+                 Background="{Binding ElementName=root, Path=CurrentBackColor}"
+                 CaretBrush="{Binding ElementName=textBox, Path=Foreground}"
+                 PreviewKeyDown="textBox_PreviewKeyDown">
+            <TextBox.Resources>
+                <Style TargetType="Border">
+                    <Setter Property="CornerRadius" Value="{Binding ElementName=root, Path=CornerRadius}"/>
+                </Style>
+            </TextBox.Resources>
+        </TextBox>
+        <TextBlock VerticalAlignment="Center"
+                   HorizontalAlignment="Left"
+                   x:Name="textBlock"
+                   Text="{Binding ElementName=root, Path=DefaultText}"
+                   Foreground="{Binding ElementName=root, Path=DefaulTextBrush}"
+                   Margin="5"
+                   MouseDown="TextBlocKMouseDown">
+        </TextBlock>
+    </Grid>
+</local:CornControl>
