+++ CornUI/CornUI.csproj
@@ -67,6 +67,9 @@
       <DependentUpon>CIcon.xaml</DependentUpon>
     </Compile>
     <Compile Include="Controls\Normal\CornControl.cs" />
+    <Compile Include="Controls\Normal\CTextbox.xaml.cs">
+      <DependentUpon>CTextbox.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Controls\ThemedCwnd.cs" />
     <Compile Include="Model\Theme.cs" />
     <Compile Include="Model\CwndInfo.cs" />
@@ -92,6 +95,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Controls\Normal\CIcon.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Controls\Normal\CTextbox.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
